+++ src/modules/editor/utils.spec.ts
@@ -1,5 +1,6 @@
 import { Locale, BodyShape, WearableCategory, WearableDefinition } from '@dcl/schemas'
-import { Color4, Wearable } from 'decentraland-ecs'
+import type { Wearable } from 'decentraland-ecs'
+import { Color4 } from '@babylonjs/core'
 import { Item, ItemType } from 'modules/item/types'
 import { getSkinHiddenCategories } from 'modules/item/utils'
 import { convertWearable, wearable, catalystWearable } from 'specs/editor'
